chahua: 拆 persona_import.py —— 抽出 persona_github / persona_provenance
1202 行单模块揉了 4 个关注点。按 DAG 拆三层、公开 API 完全不变：

- persona_provenance.py（新，~245 行）：provenance 数据模型
  （PersonaSource / GithubProvenance）+ 读写 + 内容哈希 + STATUS_* 词表
  + 基类异常 PersonaImportError。
- persona_github.py（新，~170 行）：GitHub Contents API 低层 HTTP 原语
  （_gh_get_* / _parse_github_url / _GitHubError）。
- persona_import.py（1202 → 885 行）：import / update / check / delete
  编排层 + façade re-export（含 __all__）。

DAG：persona_provenance（叶）← persona_github ← persona_import。

测试 monkeypatch persona_import._gh_latest_commit_sha / _collect_from_source
等仍生效 —— 被 patch 符号的内部调用方（_safe_latest_commit_sha / _check_github
/ _fetch_github_recursive / 生命周期函数）都留在 persona_import，经本模块命名
空间解析；persona_github 内部不自调这些符号，无旁路 call path。

行为零变化（纯结构）。全量 1371 passed。
